Trap for data items equal to 0 and conversion of some ints to long to avoid overflows on large data sets

+++ trunk/src/parallel_f.hpp
@@ -14,7 +14,7 @@ void CCPi::instrument::forward_project(const real det_y[], const real det_z[],
 				       const int ny_voxels,
 				       const int nz_voxels)
 {
-  int curr_angle, curr_ray_y, curr_ray_z;
+  long curr_angle, curr_ray_y, curr_ray_z;
   long ray_offset;
   real cos_curr_angle, sin_curr_angle;
   real start[3], end[3];
@@ -36,7 +36,8 @@ void CCPi::instrument::forward_project(const real det_y[], const real det_z[],
       cos_curr_angle = std::cos(phi[curr_angle]);
       sin_curr_angle = std::sin(phi[curr_angle]);
 
-      ray_offset = curr_angle * n_rays_y * n_rays_z + curr_ray_z*n_rays_y;
+      ray_offset = curr_angle * long(n_rays_y) * long(n_rays_z)
+	+ curr_ray_z * long(n_rays_y);
 
       /* loop over y values on detector */
       for(curr_ray_y = 0; curr_ray_y < n_rays_y; curr_ray_y++) {
